Sletter brevkode som ikke skal være i bruk lengre
