+++ src/main/kotlin/no/nav/familie/ef/sak/brev/BrevMellomlagerController.kt
@@ -1,14 +1,11 @@
 package no.nav.familie.ef.sak.brev
 
 import no.nav.familie.ef.sak.AuditLoggerEvent
-import no.nav.familie.ef.sak.brev.dto.FrittståendeBrevDto
 import no.nav.familie.ef.sak.brev.dto.MellomlagreBrevRequestDto
 import no.nav.familie.ef.sak.brev.dto.MellomlagretBrevResponse
-import no.nav.familie.ef.sak.infrastruktur.sikkerhet.SikkerhetContext
 import no.nav.familie.ef.sak.infrastruktur.sikkerhet.TilgangService
 import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.security.token.support.core.api.ProtectedWithClaims
-import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
@@ -42,32 +39,6 @@ class BrevMellomlagerController(
                 mellomlagretBrev.versjon,
             ),
         )
-    }
-
-    @Deprecated("Skal slettes")
-    @PostMapping("/frittstaende")
-    fun mellomlagreFrittstaendeBrev(@RequestBody mellomlagretBrev: FrittståendeBrevDto): Ressurs<UUID> {
-        tilgangService.validerTilgangTilFagsak(mellomlagretBrev.fagsakId, AuditLoggerEvent.UPDATE)
-        tilgangService.validerHarSaksbehandlerrolle()
-
-        return Ressurs.success(mellomlagringBrevService.mellomlagreFrittståendeBrev(mellomlagretBrev))
-    }
-
-    @Deprecated("Skal slettes")
-    @DeleteMapping("/frittstaende/{fagsakId}")
-    fun mellomlagreFrittstaendeBrev(@PathVariable fagsakId: UUID): Ressurs<UUID> {
-        tilgangService.validerTilgangTilFagsak(fagsakId, AuditLoggerEvent.DELETE)
-        tilgangService.validerHarSaksbehandlerrolle()
-        mellomlagringBrevService.slettMellomlagretFrittståendeBrev(fagsakId, SikkerhetContext.hentSaksbehandler())
-        return Ressurs.success(fagsakId)
-    }
-
-    @Deprecated("Skal slettes")
-    @GetMapping("/frittstaende/{fagsakId}")
-    fun hentMellomlagretFrittstaendeBrev(@PathVariable fagsakId: UUID): Ressurs<FrittståendeBrevDto?> {
-        tilgangService.validerTilgangTilFagsak(fagsakId, AuditLoggerEvent.ACCESS)
-        tilgangService.validerHarSaksbehandlerrolle()
-        return Ressurs.success(mellomlagringBrevService.hentMellomlagretFrittståendeBrev(fagsakId))
     }
 
     @GetMapping("/{behandlingId}")
