--- src/main/kotlin/no/nav/familie/ef/sak/brev/MellomlagerFrittståendeBrevRepository.kt
@@ -1,16 +1,0 @@
-package no.nav.familie.ef.sak.brev
-
-import no.nav.familie.ef.sak.brev.domain.MellomlagretFrittståendeBrev
-import no.nav.familie.ef.sak.repository.InsertUpdateRepository
-import no.nav.familie.ef.sak.repository.RepositoryInterface
-import org.springframework.stereotype.Repository
-import java.util.UUID
-
-@Deprecated("Skal slettes")
-@Repository
-interface MellomlagerFrittståendeBrevRepository :
-    RepositoryInterface<MellomlagretFrittståendeBrev, UUID>,
-    InsertUpdateRepository<MellomlagretFrittståendeBrev> {
-
-    fun findByFagsakIdAndSaksbehandlerIdent(fagsakId: UUID, saksehandlerIdent: String): MellomlagretFrittståendeBrev?
-}
