+++ src/main/kotlin/no/nav/familie/ef/sak/brev/MellomlagringBrevService.kt
@@ -1,12 +1,7 @@
 package no.nav.familie.ef.sak.brev
 
-import no.nav.familie.ef.sak.brev.BrevmottakerUtil.validerUnikeBrevmottakere
-import no.nav.familie.ef.sak.brev.domain.Fritekstbrev
-import no.nav.familie.ef.sak.brev.domain.FrittståendeBrevmottakere
 import no.nav.familie.ef.sak.brev.domain.MellomlagretBrev
-import no.nav.familie.ef.sak.brev.domain.MellomlagretFrittståendeBrev
 import no.nav.familie.ef.sak.brev.domain.MellomlagretFrittståendeSanitybrev
-import no.nav.familie.ef.sak.brev.dto.FrittståendeBrevDto
 import no.nav.familie.ef.sak.brev.dto.MellomlagretBrevResponse
 import no.nav.familie.ef.sak.brev.dto.MellomlagretBrevSanity
 import no.nav.familie.ef.sak.infrastruktur.sikkerhet.SikkerhetContext
@@ -18,7 +13,6 @@ import java.util.UUID
 @Service
 class MellomlagringBrevService(
     private val mellomlagerBrevRepository: MellomlagerBrevRepository,
-    private val mellomlagerFrittståendeBrevRepository: MellomlagerFrittståendeBrevRepository,
     private val mellomlagerFrittståendeSanitybrevRepository: MellomlagerFrittståendeSanitybrevRepository,
 ) {
 
@@ -48,40 +42,6 @@ class MellomlagringBrevService(
         return mellomlagerFrittståendeSanitybrevRepository.insert(mellomlagretBrev).fagsakId
     }
 
-    @Deprecated("Skal slettes")
-    fun mellomlagreFrittståendeBrev(mellomlagretBrev: FrittståendeBrevDto): UUID {
-        mellomlagretBrev.mottakere?.let { validerUnikeBrevmottakere(it) }
-        val saksbehandlerIdent = SikkerhetContext.hentSaksbehandler()
-        slettMellomlagretFrittståendeBrev(mellomlagretBrev.fagsakId, saksbehandlerIdent)
-        val mellomlagretFrittståendeBrev = MellomlagretFrittståendeBrev(
-            fagsakId = mellomlagretBrev.fagsakId,
-            brev =
-            Fritekstbrev(
-                overskrift = mellomlagretBrev.overskrift,
-                avsnitt = mellomlagretBrev.avsnitt,
-            ),
-            brevType =
-            mellomlagretBrev.brevType,
-            saksbehandlerIdent = saksbehandlerIdent,
-            mottakere = mellomlagretBrev.mottakere?.let { FrittståendeBrevmottakere(it.personer, it.organisasjoner) },
-        )
-        return mellomlagerFrittståendeBrevRepository.insert(mellomlagretFrittståendeBrev).fagsakId
-    }
-
-    @Deprecated("Skal slettes")
-    fun hentMellomlagretFrittståendeBrev(fagsakId: UUID): FrittståendeBrevDto? {
-        val saksbehandlerIdent = SikkerhetContext.hentSaksbehandler()
-        return mellomlagerFrittståendeBrevRepository.findByFagsakIdAndSaksbehandlerIdent(fagsakId, saksbehandlerIdent)?.let {
-            FrittståendeBrevDto(
-                it.brev.overskrift,
-                it.brev.avsnitt,
-                fagsakId,
-                it.brevType,
-                it.mottakere?.let { mottakere -> BrevmottakereDto(mottakere.personer, mottakere.organisasjoner) },
-            )
-        }
-    }
-
     fun hentMellomlagretFrittståendeSanitybrev(fagsakId: UUID): MellomlagretBrevResponse? =
         mellomlagerFrittståendeSanitybrevRepository.findByFagsakIdAndSaksbehandlerIdent(
             fagsakId,
@@ -103,9 +63,6 @@ class MellomlagringBrevService(
     }
 
     fun slettMellomlagretFrittståendeBrev(fagsakId: UUID, saksbehandlerIdent: String) {
-        mellomlagerFrittståendeBrevRepository.findByFagsakIdAndSaksbehandlerIdent(fagsakId, saksbehandlerIdent)
-            ?.let { mellomlagerFrittståendeBrevRepository.deleteById(it.id) }
-
         mellomlagerFrittståendeSanitybrevRepository.findByFagsakIdAndSaksbehandlerIdent(fagsakId, saksbehandlerIdent)
             ?.let { mellomlagerFrittståendeSanitybrevRepository.deleteById(it.id) }
     }
