+++ src/main/kotlin/no/nav/familie/ef/sak/brev/domain/Brevmottakere.kt
@@ -1,7 +1,9 @@
 package no.nav.familie.ef.sak.brev.domain
 
+import no.nav.familie.ef.sak.infrastruktur.sikkerhet.SikkerhetContext
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Table
+import java.time.LocalDateTime
 import java.util.UUID
 
 @Table
@@ -27,3 +29,14 @@ data class BrevmottakerOrganisasjon(
 
 data class PersonerWrapper(val personer: List<BrevmottakerPerson>)
 data class OrganisasjonerWrapper(val organisasjoner: List<BrevmottakerOrganisasjon>)
+
+@Table("brevmottakere_frittstaende_brev")
+data class BrevmottakereFrittståendeBrev(
+    @Id
+    val id: UUID = UUID.randomUUID(),
+    val fagsakId: UUID,
+    val saksbehandlerIdent: String = SikkerhetContext.hentSaksbehandler(),
+    val tidspunktOpprettet: LocalDateTime = LocalDateTime.now(),
+    val personer: PersonerWrapper,
+    val organisasjoner: OrganisasjonerWrapper,
+)
