--- src/main/kotlin/no/nav/familie/ef/sak/brev/domain/Fritekstbrev.kt
@@ -1,5 +1,0 @@
-package no.nav.familie.ef.sak.brev.domain
-
-import no.nav.familie.ef.sak.brev.dto.FrittståendeBrevAvsnitt
-
-data class Fritekstbrev(val overskrift: String, val avsnitt: List<FrittståendeBrevAvsnitt>)
