--- src/main/kotlin/no/nav/familie/ef/sak/brev/dto/FritekstBrevKategori.kt
@@ -1,17 +1,0 @@
-package no.nav.familie.ef.sak.brev.dto
-
-@Deprecated("Skal slettes")
-enum class FritekstBrevKategori {
-    SANKSJON,
-    VEDTAK_INVILGELSE,
-    VEDTAK_AVSLAG,
-    VEDTAK_OPPHØR,
-    VEDTAK_REVURDERING,
-    VEDTAK_INNVILGELSE_BARNETILSYN,
-    VEDTAK_AVSLAG_BARNETILSYN,
-    VEDTAK_INNVILGELSE_SKOLEPENGER,
-    VEDTAK_AVSLAG_SKOLEPENGER,
-    VEDTAK_OPPHØR_BARNETILSYN,
-    VEDTAK_OPPHØR_SKOLEPENGER,
-    VEDTAK_UTESTENGELSE,
-}
