--- src/main/kotlin/no/nav/familie/ef/sak/brev/dto/FrittståendeBrevRequestDto.kt
@@ -1,9 +1,0 @@
-package no.nav.familie.ef.sak.brev.dto
-
-@Deprecated("Skal slettes")
-data class FrittståendeBrevRequestDto(
-    val overskrift: String,
-    val avsnitt: List<FrittståendeBrevAvsnitt>,
-    val personIdent: String,
-    val navn: String,
-)
