+++ src/main/kotlin/no/nav/familie/ef/sak/brev/dto/MellomlagretBrevResponse.kt
@@ -1,19 +1,8 @@
 package no.nav.familie.ef.sak.brev.dto
 
-import no.nav.familie.ef.sak.brev.domain.Fritekstbrev
-
 sealed class MellomlagretBrevResponse
 
 data class MellomlagretBrevSanity(val brevtype: Brevtype = Brevtype.SANITYBREV, val brevmal: String, val brevverdier: String?) :
-    MellomlagretBrevResponse()
-
-@Deprecated("Skal slettes")
-data class MellomlagretBrevFritekst(
-    val brevtype: Brevtype = Brevtype.FRITEKSTBREV,
-    val brev: Fritekstbrev,
-    val brevType: FritekstBrevKategori,
-    val brevmal: String = "Fritekstbrev",
-) :
     MellomlagretBrevResponse()
 
 enum class Brevtype {
