--- src/main/kotlin/no/nav/familie/ef/sak/brev/dto/VedtaksbrevFritekstDto.kt
@@ -1,10 +1,0 @@
-package no.nav.familie.ef.sak.brev.dto
-
-import java.util.UUID
-
-@Deprecated("Skal slettes")
-data class VedtaksbrevFritekstDto(
-    val overskrift: String,
-    val avsnitt: List<FrittståendeBrevAvsnitt>,
-    val behandlingId: UUID,
-)
