+++ src/main/kotlin/no/nav/familie/ef/sak/infrastruktur/config/DatabaseConfiguration.kt
@@ -1,8 +1,6 @@
 package no.nav.familie.ef.sak.infrastruktur.config
 
 import com.fasterxml.jackson.module.kotlin.readValue
-import no.nav.familie.ef.sak.brev.domain.Fritekstbrev
-import no.nav.familie.ef.sak.brev.domain.FrittståendeBrevmottakere
 import no.nav.familie.ef.sak.brev.domain.OrganisasjonerWrapper
 import no.nav.familie.ef.sak.brev.domain.PersonerWrapper
 import no.nav.familie.ef.sak.felles.domain.Endret
@@ -127,16 +125,12 @@ class DatabaseConfiguration : AbstractJdbcConfiguration() {
                 DetaljertSimuleringResultatTilPGobjectConverter(),
                 PGobjectTilBeriketSimuleringsresultat(),
                 BeriketSimuleringsresultatTilPGobjectConverter(),
-                PGObjectTilFritekstbrevConverter(),
-                FritekstbrevTilPGObjectConverter(),
                 PGobjectTilBrevmottakerPersoner(),
                 BrevmottakerePersonerTilPGobjectConverter(),
                 PGobjectTilBrevmottakerOrganisasjoner(),
                 BrevmottakereOrganisasjonerTilPGobjectConverter(),
                 PGobjectTilSkolepengerConverter(),
                 SkolepengerTilPGobjectConverter(),
-                PGobjectTilFrittståendeBrevmottakere(),
-                FrittståendeBrevmottakereTilPGobjectConverter(),
             ),
         )
     }
@@ -444,24 +438,6 @@ class DatabaseConfiguration : AbstractJdbcConfiguration() {
     }
 
     @ReadingConverter
-    class PGObjectTilFritekstbrevConverter : Converter<PGobject, Fritekstbrev?> {
-
-        override fun convert(pGobject: PGobject): Fritekstbrev? {
-            return pGobject.value?.let { objectMapper.readValue(it) }
-        }
-    }
-
-    @WritingConverter
-    class FritekstbrevTilPGObjectConverter : Converter<Fritekstbrev, PGobject> {
-
-        override fun convert(simuleringsresultat: Fritekstbrev): PGobject =
-            PGobject().apply {
-                type = "json"
-                value = objectMapper.writeValueAsString(simuleringsresultat)
-            }
-    }
-
-    @ReadingConverter
     class PGobjectTilBrevmottakerPersoner : Converter<PGobject, PersonerWrapper> {
 
         override fun convert(pGobject: PGobject): PersonerWrapper? {
@@ -491,24 +467,6 @@ class DatabaseConfiguration : AbstractJdbcConfiguration() {
     class BrevmottakereOrganisasjonerTilPGobjectConverter : Converter<OrganisasjonerWrapper, PGobject> {
 
         override fun convert(mottakere: OrganisasjonerWrapper): PGobject =
-            PGobject().apply {
-                type = "json"
-                value = objectMapper.writeValueAsString(mottakere)
-            }
-    }
-
-    @ReadingConverter
-    class PGobjectTilFrittståendeBrevmottakere : Converter<PGobject, FrittståendeBrevmottakere> {
-
-        override fun convert(pGobject: PGobject): FrittståendeBrevmottakere? {
-            return pGobject.value?.let { objectMapper.readValue(it) }
-        }
-    }
-
-    @WritingConverter
-    class FrittståendeBrevmottakereTilPGobjectConverter : Converter<FrittståendeBrevmottakere, PGobject> {
-
-        override fun convert(mottakere: FrittståendeBrevmottakere): PGobject =
             PGobject().apply {
                 type = "json"
                 value = objectMapper.writeValueAsString(mottakere)
