+++ src/main/resources/db/migration/V140__slett_mellomlager_frittstående_brev.sql
@@ -1,0 +1,2 @@
+DROP TABLE mellomlagret_frittstaende_brev;
+
